feature: SpringFactories中AOT分离出来单独处理
