+++ docs/asciidoctor/spring-auto-service.adoc
@@ -74,8 +74,6 @@ com.livk.commons.http.WebClientConfiguration
 
 ==== @SpringFactories 使用示例
 
-@SpringFactories支持生成aot.factories原理基本同下，只需指定属性aot=true
-
 指定接口为spring.factories的Key
 
 [source,java,indent=0]
@@ -105,6 +103,44 @@ public class TraceEnvironmentPostProcessor implements EnvironmentPostProcessor {
 ----
 
 生成文件 META-INF/spring.factories
+
+[source,properties,indent=0]
+----
+org.springframework.boot.env.EnvironmentPostProcessor=\
+	com.livk.commons.spring.TraceEnvironmentPostProcessor
+----
+
+==== @AotFactories 使用示例
+
+指定接口为aot.factories的Key
+
+[source,java,indent=0]
+----
+@AotFactories(org.springframework.boot.env.EnvironmentPostProcessor)
+public class TraceEnvironmentPostProcessor implements EnvironmentPostProcessor {
+
+}
+----
+
+生成文件 META-INF/spring/aot.factories
+
+[source,properties,indent=0]
+----
+org.springframework.boot.env.EnvironmentPostProcessor=\
+	com.livk.commons.spring.TraceEnvironmentPostProcessor
+----
+
+当前类如果仅仅只有一个接口，可以不指定，自动生成
+
+[source,java,indent=0]
+----
+@AotFactories
+public class TraceEnvironmentPostProcessor implements EnvironmentPostProcessor {
+
+}
+----
+
+生成文件 META-INF/spring/aot.factories
 
 [source,properties,indent=0]
 ----
