+++ spring-auto-service/src/main/java/com/livk/auto/service/annotation/SpringFactories.java
@@ -43,7 +43,9 @@ public @interface SpringFactories {
 	/**
 	 * 将spring.factories文件调整至aot.factories
 	 * @return the boolean
+	 * @deprecated use {@link AotFactories}
 	 */
+	@Deprecated(since = "1.4.5")
 	boolean aot() default false;
 
 }
