+++ spring-auto-service/src/main/resources/META-INF/gradle/incremental.annotation.processors
@@ -1,2 +1,3 @@
 com.livk.auto.service.processor.SpringAutoServiceProcessor,AGGREGATING
 com.livk.auto.service.processor.SpringFactoriesProcessor,AGGREGATING
+com.livk.auto.service.processor.AotFactoriesProcessor,AGGREGATING
