+++ spring-auto-service/src/test/java/com/livk/auto/service/processor/AotFactoriesProcessorTest.java
@@ -1,24 +1,9 @@
-/*
- * Copyright 2021-2024 spring-boot-extension the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package com.livk.auto.service.processor;
 
-import com.livk.auto.service.processor.factories.spring.SpringFactoryService;
+import com.livk.auto.service.processor.factories.aot.AotRuntimeHintsRegistrar;
 import com.livk.auto.service.processor.factories.spring.SpringFactoryServiceImpl;
 import org.junit.jupiter.api.Test;
+import org.springframework.aot.hint.RuntimeHintsRegistrar;
 import org.springframework.core.io.UrlResource;
 import org.springframework.core.test.tools.SourceFile;
 import org.springframework.core.test.tools.TestCompiler;
@@ -36,22 +21,21 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 /**
  * @author livk
  */
-
-class SpringFactoriesProcessorTest {
+public class AotFactoriesProcessorTest {
 
 	@Test
 	void test() {
-		compile(SpringFactoryServiceImpl.class, SpringFactoryService.class);
+		compile(AotRuntimeHintsRegistrar.class, RuntimeHintsRegistrar.class);
 	}
 
 	private void compile(Class<?> type, Class<?> factoryClass) {
 		TestCompiler.forSystem()
-			.withProcessors(new SpringFactoriesProcessor())
+			.withProcessors(new AotFactoriesProcessor())
 			.withSources(SourceFile.forTestClass(type))
 			.compile(compiled -> {
 				try {
 					Enumeration<URL> resources = compiled.getClassLoader()
-						.getResources(SpringFactoriesProcessor.SPRING_LOCATION);
+						.getResources(AotFactoriesProcessor.AOT_LOCATION);
 					Properties pro = new Properties();
 					for (URL url : Collections.list(resources)) {
 						InputStream inputStream = new UrlResource(url).getInputStream();
