--- spring-auto-service/src/test/java/com/livk/auto/service/processor/SpringFactoryServiceImpl.java
@@ -1,29 +1,0 @@
-/*
- * Copyright 2021-2024 spring-boot-extension the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.livk.auto.service.processor;
-
-import com.google.auto.service.AutoService;
-import com.livk.auto.service.annotation.SpringFactories;
-
-/**
- * @author livk
- */
-@SpringFactories
-@AutoService(SpringFactoryService.class)
-public class SpringFactoryServiceImpl implements SpringFactoryService {
-
-}
