+++ spring-auto-service/src/test/java/com/livk/auto/service/processor/autoconfigure/EnableAuto.java
@@ -1,0 +1,15 @@
+package com.livk.auto.service.processor.autoconfigure;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * @author livk
+ */
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface EnableAuto {
+
+}
