+++ spring-auto-service/src/test/java/com/livk/auto/service/processor/autoconfigure/SpringAutoContext.java
@@ -1,0 +1,11 @@
+package com.livk.auto.service.processor.autoconfigure;
+
+import com.livk.auto.service.annotation.SpringAutoService;
+
+/**
+ * @author livk
+ */
+@SpringAutoService(EnableAuto.class)
+public class SpringAutoContext {
+
+}
